% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{make_names}
\alias{make_names}
\title{Make column names for a onehot object}
\usage{
make_names(x)
}
\arguments{
\item{x}{a \code{\link{onehot}} object}
}
\description{
Make column names for a onehot object
}
\examples{
data(iris)
encoder <- onehot(iris)
make_names(encoder)
}
