% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.onehot}
\alias{predict.onehot}
\title{Predict onehot objects}
\usage{
\method{predict}{onehot}(object, data, sparse = FALSE, sep = "_", ...)
}
\arguments{
\item{object}{an object of class \code{\link{onehot}}}

\item{data}{a data.frame to onehot encode useing \code{object}}

\item{sparse}{if TRUE, returns a \code{\link[Matrix]{dgCMatrix-class}}}

\item{...}{further arguments passed to or from other methods}
}
\value{
a matrix with factor variable onehot encoded
}
\description{
Predict onehot objects
}
\examples{
data(iris)
encoder <- onehot(iris)
x <- predict(encoder, iris)
x_sparse <- predict(encoder, iris, sparse=TRUE)
}
