% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onehot.R
\name{print.onehot}
\alias{print.onehot}
\title{Print information about onehot object}
\usage{
\method{print}{onehot}(x, ...)
}
\arguments{
\item{x}{onehot object}

\item{...}{further arguments passed to or from other methods}
}
\description{
Print information about onehot object
}

