% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onehot.R
\name{summary.onehot}
\alias{summary.onehot}
\title{Summarize onehot object}
\usage{
\method{summary}{onehot}(object, ...)
}
\arguments{
\item{object}{onehot object to summarize}

\item{...}{further arguments passed to or from other methods}
}
\description{
Summarize onehot object
}

