% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onehot.R
\name{onehot}
\alias{onehot}
\title{Onehot encode a data.frame}
\usage{
onehot(data, sentinel = .Machine$double.xmin, stringsAsFactors = FALSE,
  max_levels = 10)
}
\arguments{
\item{data}{data.frame to convert factors into onehot encoded columns}

\item{stringsAsFactors}{if TRUE, converts character vectors to factors}

\item{max_levels}{maximum number of levels to onehot encode per factor
variable. Factors with levels exceeding this number will be skipped.}

\item{addNA}{if TRUE, all variables get an NA indicator}
}
\value{
a \code{onehot} object descrbing how to transform the data
}
\description{
Onehot encode a data.frame
}
\details{
By default, with \code{addNA=FALSE}, no NAs are returned for
non-factor columns. Indicator columns are created for factor levels and NA
factors are ignored. The exception is when NA is an explicit factor level.

\code{stringsAsFactrs=TRUE} will convert character columns to factors first.
Other wise characters are ignored. Only factor, numeric, integer, and logical
vectors are valid for onehot. Other classes will be skipped entirely.

\code{addNA=TRUE} will create indicator columns for every field. This will
add ncols columns to the output matrix. A sparse matrix may be better in
such cases.
}
\examples{
data(iris)
encoder <- onehot(iris)

## add NA indicator columns
encoder <- onehot(iris, addNA=TRUE)

## Convert character fields to factrs
encoder <- onehot(iris, stringsAsFactors=TRUE)

## limit which factors are onehot encoded
encoder <- onehot(iris, max_levels=5)
}
