% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onehot.R
\name{print.onehot}
\alias{print.onehot}
\title{Print information about a onehot object}
\usage{
\method{print}{onehot}(x, ...)
}
\arguments{
\item{x}{onehot object to print}

\item{...}{other arguments pass to or from other functions}
}
\description{
Print information about a onehot object
}
