% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{summary.onehot}
\alias{summary.onehot}
\title{Summarize onehot object}
\usage{
\method{summary}{onehot}(object, ...)
}
\arguments{
\item{object}{a onehot object}

\item{...}{other arguments pass to or from other functions}
}
\description{
Summarize onehot object
}
\examples{
## Create some dummy data with different column types
x <- data.frame(HairEyeColor)
x$Hair <- as.character(x$Hair)

## Create a onehot object
encoder <- onehot(x)

## Return a list with summary information
summary(encoder)
}
